fix(api): consolidate follow/unfollow API into a userRepo
