+++ src/users/infrastructure/persistence/user.abstract.repository.ts
@@ -1,6 +1,7 @@
 import { SelectQueryBuilder } from 'typeorm';
 
 import { User } from '@src/users/domain/user';
+import { UserFollow } from '@src/users/domain/user-follow';
 import { FilterUserDto, SortUserDto } from '@src/users/dto/query-user.dto';
 import { DeepPartial } from '@src/utils/types/deep-partial.type';
 import { NullableType } from '@src/utils/types/nullable.type';
@@ -51,4 +52,15 @@ export abstract class UserAbstractRepository {
     id: User['id'],
     query: SelectQueryBuilder<UserSummaryViewEntity>,
   ): Promise<NullableType<UserSummary>>;
+
+  abstract createFollow(
+    data: Omit<UserFollow, 'id' | 'createdAt' | 'updatedAt'>,
+  ): Promise<UserFollow>;
+
+  abstract findFollow(
+    followerId: User['id'],
+    followingId: User['id'],
+  ): Promise<NullableType<UserFollow>>;
+
+  abstract removeFollow(id: UserFollow['id']): Promise<void>;
 }
